requirements.txt
alabaster==1.0.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
babel==2.18.0
black==26.1.0
blinker==1.9.0
causalgraphicalmodels==0.0.4
certifi==2026.1.4
cffi==1.15.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cplex==22.1.2.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dash==3.4.0
decorator==5.2.1
dice_ml==0.12
distlib==0.4.0
docutils==0.22.4
executing==2.2.1
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
graphviz==0.21
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
identify==2.6.16
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
importlib_metadata==8.7.1
iniconfig==2.3.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightgbm==4.6.0
lime==0.2.0.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mip==1.12.0
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.2
numpydoc==1.10.0
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21
# nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.5
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvshmem-cu12==3.4.5
# nvidia-nvtx-cu12==12.8.90
pandas==3.0.0
parso==0.8.5
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
pre_commit==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.2
PySMT==0.9.6
pytest==9.0.2
python-dateutil==2.9.0.post0
# python==3.12.12
pytokens==0.4.1
PyYAML==6.0.3
raiutils==0.4.2
recourse==1.0.0
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
roman-numerals==4.1.0
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinx-autodoc-typehints==3.6.2
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.1.28
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.10.0+cu126
torchvision==0.25.0+cu126
tqdm==4.67.2
# traitlets==5.14.3
# triton==3.6.0
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
wcwidth==0.5.3
Werkzeug==3.1.5
wheel==0.46.3
xgboost==3.1.3
zipp==3.23.0
z3-solver==4.16.0.0